minor refactoring and docs
